conditionals/each: Integrate with TupleState, clean up other conditionals
