+++ src/extensions/react/conditionals/switch.ts
@@ -1,5 +1,4 @@
-// TODO: Do we want to compute these?
-import { useComputed } from "implementations/react/hooks";
+// TODO: Do we want to compute/memoize these?
 import type { ReactNode } from "react";
 
 export type SwitchProps<T extends string | symbol | number = any> = {
@@ -13,7 +12,7 @@ export type SwitchProps<T extends string | symbol | number = any> = {
 export function Switch<T extends string | symbol | number = any>({
     q, when
 }: SwitchProps<T>) {
-    const value = useComputed(q).use();
+    const value = q();
 
     return Object.hasOwn(when, value) ? when[value] : when.default;
 }
